role3_perception/requirements.txt
numpy<2
pillow
ultralytics
pytest
